Refactor last re-org tests
